Moved plugin settings into their own file
